feat(settings): Add vercel log drain endpoint to project key settings
