+++ static/app/views/settings/project/loaderScript.spec.tsx
@@ -100,6 +100,7 @@ describe('LoaderScript', () => {
           crons: '',
           playstation:
             'http://dev.getsentry.net:8000/api/1/playstation?sentry_key=188ee45a58094d939428d8585aa6f662',
+          integration: 'http://dev.getsentry.net:8000/api/1/integration/',
           otlp_traces: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/traces',
           otlp_logs: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/logs',
         },
@@ -242,6 +243,7 @@ describe('LoaderScript', () => {
           crons: '',
           playstation:
             'http://dev.getsentry.net:8000/api/1/playstation?sentry_key=188ee45a58094d939428d8585aa6f662',
+          integration: 'http://dev.getsentry.net:8000/api/1/integration/',
           otlp_traces: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/traces',
           otlp_logs: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/logs',
         },
