+++ static/app/views/settings/project/projectKeys/list/index.spec.tsx
@@ -176,6 +176,7 @@ describe('ProjectKeys', () => {
           crons: '',
           playstation:
             'http://dev.getsentry.net:8000/api/1/playstation?sentry_key=188ee45a58094d939428d8585aa6f662',
+          integration: 'http://dev.getsentry.net:8000/api/1/integration/',
           otlp_traces: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/traces',
           otlp_logs: 'http://dev.getsentry.net:8000/api/1/integration/otlp/v1/logs',
         },
